requirements.txt
altgraph==0.17.4
packaging==23.2
pefile==2023.2.7
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
pyserial==3.5
pywin32-ctypes==0.2.2
